Refactor .h/.cpp (except ConvertLayout)
