+++ lib/Conversion/TritonGPUToLLVM/ViewOpToLLVM.h
@@ -6,99 +6,6 @@
 using namespace mlir;
 using namespace mlir::triton;
 
-using ::mlir::LLVM::getElementsFromStruct;
-using ::mlir::LLVM::getStructFromElements;
-using ::mlir::triton::gpu::getElemsPerThread;
-
-struct SplatOpConversion
-    : public ConvertTritonGPUOpToLLVMPattern<triton::SplatOp> {
-  using ConvertTritonGPUOpToLLVMPattern<
-      triton::SplatOp>::ConvertTritonGPUOpToLLVMPattern;
-
-  // Convert SplatOp or arith::ConstantOp with SplatElementsAttr to a
-  // LLVM::StructType value.
-  //
-  // @elemType: the element type in operand.
-  // @resType: the return type of the Splat-like op.
-  // @constVal: a LLVM::ConstantOp or other scalar value.
-  static Value convertSplatLikeOp(Type elemType, Type resType, Value constVal,
-                                  TypeConverter *typeConverter,
-                                  ConversionPatternRewriter &rewriter,
-                                  Location loc);
-
-  static Value convertSplatLikeOpWithMmaLayout(const MmaEncodingAttr &layout,
-                                               Type resType, Type elemType,
-                                               Value constVal,
-                                               TypeConverter *typeConverter,
-                                               ConversionPatternRewriter &rewriter,
-                                               Location loc);
-
-  LogicalResult
-  matchAndRewrite(triton::SplatOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const;
-};
-
-// This pattern helps to convert arith::ConstantOp(with SplatElementsAttr),
-// the logic is the same as triton::SplatOp, so the underlying implementation
-// is reused.
-struct ArithConstantSplatOpConversion
-    : public ConvertTritonGPUOpToLLVMPattern<arith::ConstantOp> {
-  using ConvertTritonGPUOpToLLVMPattern<
-      arith::ConstantOp>::ConvertTritonGPUOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(arith::ConstantOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override;
-};
-
-struct CatOpConversion : public ConvertTritonGPUOpToLLVMPattern<CatOp> {
-  using OpAdaptor = typename CatOp::Adaptor;
-
-  explicit CatOpConversion(LLVMTypeConverter &typeConverter,
-                           PatternBenefit benefit = 1)
-      : ConvertTritonGPUOpToLLVMPattern<CatOp>(typeConverter, benefit) {}
-
-  LogicalResult
-  matchAndRewrite(CatOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override;
-};
-
-template <typename SourceOp>
-struct ViewLikeOpConversion : public ConvertTritonGPUOpToLLVMPattern<SourceOp> {
-  using OpAdaptor = typename SourceOp::Adaptor;
-  explicit ViewLikeOpConversion(LLVMTypeConverter &typeConverter,
-                                PatternBenefit benefit = 1)
-      : ConvertTritonGPUOpToLLVMPattern<SourceOp>(typeConverter, benefit) {}
-
-  LogicalResult
-  matchAndRewrite(SourceOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override {
-    // We cannot directly run `rewriter.replaceOp(op, adaptor.src())`
-    // due to MLIR's restrictions
-    Location loc = op->getLoc();
-    auto resultTy = op.getType().template cast<RankedTensorType>();
-    unsigned elems = getElemsPerThread(resultTy);
-    Type elemTy =
-        this->getTypeConverter()->convertType(resultTy.getElementType());
-    SmallVector<Type> types(elems, elemTy);
-    Type structTy = LLVM::LLVMStructType::getLiteral(this->getContext(), types);
-    auto vals = getElementsFromStruct(loc, adaptor.src(), rewriter);
-    Value view = getStructFromElements(loc, vals, rewriter, structTy);
-    rewriter.replaceOp(op, view);
-    return success();
-  }
-};
-
-struct TransOpConversion
-    : public ConvertTritonGPUOpToLLVMPattern<triton::TransOp> {
-  using ConvertTritonGPUOpToLLVMPattern<
-      triton::TransOp>::ConvertTritonGPUOpToLLVMPattern;
-
-  LogicalResult
-  matchAndRewrite(triton::TransOp op, OpAdaptor adaptor,
-                  ConversionPatternRewriter &rewriter) const override;
-};
-
 void populateViewOpToLLVMPatterns(
     mlir::LLVMTypeConverter &typeConverter,
     RewritePatternSet &patterns, int numWarps,
